# PoC: Refactor messaging system with PSR container interface