… composer.json
… packages/Ecotone/src/Messaging/Config/MessagingSystemConfiguration.php
… packages/Ecotone/src/Messaging/Config/PreparedConfiguration.php
… packages/Ecotone/src/Messaging/Endpoint/EventDriven/EventDrivenConsumer.php
… packages/Ecotone/src/Messaging/Endpoint/InterceptedConsumer.php
… packages/Ecotone/src/Messaging/Handler/ChannelResolver.php
… packages/Ecotone/src/Messaging/Handler/Gateway/ProxyFactory.php
… packages/Ecotone/src/Messaging/Handler/InterfaceToCallRegistry.php
… packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/AroundInterceptorReference.php
… packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/MethodCall.php
… packages/Ecotone/src/Messaging/Handler/Processor/MethodInvoker/MethodInvoker.php
… packages/Ecotone/src/Messaging/Handler/RequestReplyProducer.php
… packages/Ecotone/src/Messaging/Handler/ServiceProviderInterface.php
… packages/Symfony/App/Kernel.php
… packages/Symfony/CacheWarmer/ProxyCacheWarmer.php
… packages/Symfony/DepedencyInjection/Compiler/CacheCleaner.php
… packages/Symfony/DepedencyInjection/Compiler/EcotoneCompilerPass.php
… packages/Symfony/DepedencyInjection/Configuration.php
… packages/Symfony/DepedencyInjection/EcotoneExtension.php
… packages/Symfony/DepedencyInjection/MessagingEntrypointCommand.php
… packages/Symfony/SymfonyBundle/EcotoneSymfonyBundle.php
… packages/Symfony/SymfonyBundle/MessagingSystemFactory.php
… packages/Symfony/SymfonyBundle/PreparedConfigurationFromDumpFactory.php
… packages/Symfony/SymfonyBundle/Proxy/Autoloader.php
… packages/Symfony/SymfonyBundle/Proxy/ProxyFactory.php
… packages/Symfony/composer.json
… packages/Symfony/config/routes/annotations.yaml
… packages/Symfony/tests/App/Kernel.php
… packages/Symfony/tests/App/bin/console
… packages/Symfony/tests/App/config/bootstrap.php