Many different things to discuss in this PR.
The first goal was to lazy load every messaging components (gateway, channels, etc.) thanks to already existing container implementation, and this PR uses exclusively Symfony for the proof of concept. It is related to issue #133 .

From my understanding, Ecotone is built around:

- `MessagingSystemConfiguration`: where all messaging components are configured, but I think not as much as it should (some "Builders" leak elsewhere in the code). This configuration is dumped for performance. I think this component looks a lot like a `ContainerBuilder` in PHP-DI or Symfony world. 
- `MessagingSystem`: where all built components are registered (and build in the static function `buildFromMessagingConfiguration` but that is just moving code around). This component looks a lot like a `Container`
- `Modules`: where you can hook into the `MessagingSystemConfiguration` (the `ContainerBuilder`), this looks a lot like a `Bundle` in Symfony world.

I think it could be possible for Ecotone to rely on external `ContainerBuilder` and `Container` implementations and propose a default one in `EcotoneLite`. This could lead to a lot of simplification and optimization of Ecotone.
A common interface could be that all "Builders" have a `build` method that accept a single parameter `EcotoneContainerBuilder`. Instead of returning an actual object, "Builders" build method could configure the `EcotoneContainerBuilder` on *how* to build the object using dependency injection, like a ContainerBuilder in Symfony of PHP-DI.

The downside is that there is a necessity to maintain multiple container implementation (Symfony/Laravel/Lite), and as we rely more heavily on them, we may have to run all test suites on each container implementation.

Another downside is that it is not compatible with EcotoneLite testing framework which rely on replacing the `ConfiguredMessagingSystem` inside the provided container (the replaceWith method). I think this is not a problem, as the `replaceWith` method is bit of a hack to me. Moreover, I think it is a strange to propose to build a new container for each test case as it is with EcotoneLite (the `$containerOrAvailableServices` parameter). It makes sense for testing Ecotone, but for a real world app, you have a single container in production, and you want to make all your tests against this one or something very near from it. And not configure a specific container for each test case. If it makes sense for some unit tests , we could use EcotoneLite without any configured container, so the problem is solved.

While building a container with Ecotone, it also required pushing a maximum of code that could be done only at compilation time vs code that should be dynamic (that depends on the `Message`), improving performance.
There is quite a big impact on performance with this PR when opcache is enabled, mainly due to how the configuration is dumped as a php file instead of using `unserialize`. But there are other perf improvements.
Here are my figures (you should look only at Symfony which implements the new system) which shows a x3 performance boost:

This PR (opcache enabled):

| benchmark           | subject            | revs | its | mem_peak | mode      | rstdev |
| ------------------- | -------------------| ---- | --- | -------- | --------- | ------ |
| PlaceOrderBenchmark | bench_symfony_prod | 10   | 5   | 5.069mb  | 763.504μs | ±2.08% |
| PlaceOrderBenchmark | bench_symfony_dev  | 10   | 5   | 5.166mb  | 798.032μs | ±0.71% |
| PlaceOrderBenchmark | bench_laravel_prod | 10   | 5   | 8.869mb  | 68.192ms  | ±1.07% |
| PlaceOrderBenchmark | bench_laravel_dev  | 10   | 5   | 8.869mb  | 69.248ms  | ±1.61% |
| PlaceOrderBenchmark | bench_lite_prod    | 10   | 5   | 12.879mb | 3.503ms   | ±0.42% |
| PlaceOrderBenchmark | bench_lite_dev     | 10   | 5   | 7.717mb  | 69.470ms  | ±1.75% |

Main (opcache enabled):
| benchmark           | subject            | revs | its | mem_peak | mode     | rstdev |
| --------------------| ------------------ | ---- | --- | -------- | -------- | ------ |
| PlaceOrderBenchmark | bench_symfony_prod | 10   | 5   | 13.148mb | 2.310ms  | ±0.45% |
| PlaceOrderBenchmark | bench_symfony_dev  | 10   | 5   | 13.261mb | 2.284ms  | ±1.18% |
| PlaceOrderBenchmark | bench_laravel_prod | 10   | 5   | 8.936mb  | 69.296ms | ±1.64% |
| PlaceOrderBenchmark | bench_laravel_dev  | 10   | 5   | 8.936mb  | 69.075ms | ±0.94% |
| PlaceOrderBenchmark | bench_lite_prod    | 10   | 5   | 12.822mb | 3.289ms  | ±2.71% |
| PlaceOrderBenchmark | bench_lite_dev     | 10   | 5   | 7.779mb  | 68.737ms | ±0.72% |

I think this PR is should be regarded as a proof of concept, but we can peak some features not directly related to the Container refactoring thing. Here are the one i am thinking of:
- Dumping a `PreparedConfiguration` object as a php file for opcache optimization (using VarExporter component that is already included with ecotone symfony bundle)
- Using a simpler code generation for proxies without proxymanager package but using directly laminas-code, because we are just proxyfying interfaces only which is simple (see packages/Symfony/SymfonyBundle/Proxy/ProxyFactory.php and an example snapshot of a proxy: packages/Symfony/tests/phpunit/ProxyFactoryTest.snapshot)
- Symfony bundle is now unregistering the autoloader on kernel shutdown, which removes a memory leak. We may consider having a shutdown function in EcotoneLite to unregister the autoloader.
- Transform the `MessagingSystem` to a simple container. The only thing that is done in the `MessagingSystem` is to `run` event driven consumers. This PR makes event driven consumer subscribe when its input channel is loaded.