@dgafka 2023-06-16
I will find time to review this on saturday/sunday, as this is interesting idea and I want to spend some time on rethinking this.