@dgafka 2023-06-16
btw the times with opcache are impressive, it's like we would simply execute clear PHP code.