@dgafka 2023-06-18
I would love to see separate PRs or discuss those in separate topics, as there is a lot of going on here :smile: 
This makes is a bit hard to follow, as I am not sure what change references to what topic. 
I've started [Github Discussions in ecotone-dev](https://github.com/ecotoneframework/ecotone-dev/discussions), so if you want to discuss given topic we can also do it there from now.
But let's try to keep one topic at time, as the changes you proposed can be separately discussed and merge regardless of each other :)

Giving feedback accordingly to what I learned from the PR:

- Using var exporter
I would love to see this as separate PR. I also wonder, what are the results without the opcache for this one. 
If there is general improvement in booting Ecotone with and without opcache I think we can make it part of core Ecotone package and replace native php serialization.
- Using laminas-code to generate proxies
What is the reason for maintaining this, does it the speed run time somehow? If so I am happy to see that as separate PR
- Autoloader registered and deregistered only when needed
Yep would love to see a separate PR working for Lite and Laravel too :)
- Subscribe event driven consumer subscribe when its input channel is loaded.
I would also love to see that as separate PR. I agree with the approach of executing only the components that are necessary in given flow

---

> It makes sense for testing Ecotone, but for a real world app, you have a single container in production, and you want to make all your tests against this one or something very near from it. And not configure a specific container for each test case. If it makes sense for some unit tests , we could use EcotoneLite without any configured container, so the problem is solved.

Well this is one is actually challenging. If you consider that you want to reuse the configuration from your Symfony App, then you will bring configuration of all components with you.
The reason to boot new Ecotone Lite instance is to bring context specific classes to the scenario without any additional stuff.
This allows for testing components in full isolation, like only aggregate, which does not trigger Event Handlers (because they are not even registered), switch from synchronous to asynchronous code, providing stubs on the places where it does real integration etc.
You can still use Ecotone directly from Symfony DI Container by fetching and executing Command Bus. Personally I actually found it low value because the context is just too big and tests on Ecotone Lite are actually near to the production usage.

If you feel like you there is value in this, let's discuss separately on github discussions.

> While building a container with Ecotone, it also required pushing a maximum of code that could be done only at compilation time vs code that should be dynamic (that depends on the Message), improving performance.

I fully agree. Ideal is that the execution code, executes and builds only Message Handlers/Channels that are being triggered.
There is a lot of potential to improve thanks to decoupled nature of the components.

> I think it could be possible for Ecotone to rely on external ContainerBuilder and Container implementations and propose a default one in EcotoneLite

If you take a look on how Symfony or Laravel are used, there is minimalistic amount of the code, even Console Commands are done on Ecotone side and just re-registered in Symfony/Laravel.
The reason for this is maintainece and higher assurence that testing on lower level components will also work in higher level environments. 
There is a lot of cool projects that failed on this, including Prooph which had really hard time maintaining the changes that were done in Symfony and keeping it up to date.
With current approach, new versions of Symfony or Laravel are not really a problem, as the space of the integration is really small. Thanks to that there is high assurance that new versions of those framework will work without any or small amount of changes.

That's why there would need to be really strong argument for pushing this outside of Ecotone as current implementation is for a reason :) 
If you want we can discuss this one separately too.

---

I am really glad that you want to tackle so many improvements and I see a lot of potential in those. This may actually push Ecotone with performance to new level, thank you :)