@jlabedo 2023-06-19
Thank you for taking time reviewing this. I was really throwing some ideas while understanding how ecotone is built and I guess it was not an easy one to review! I think we can close this PR and discuss the main subject of, let's say, "Messaging System compilation" in a github discussion as you suggest.

Regarding proxy generation, I dropped proxymanager and used laminas-code directly (which is also used by proxymanager) because I had difficulties understanding some parts of proxying code. I can remember of:
- the warmup phase that was doing strange things to me (like warming an empty proxy or something like that ?)
- the proxy Configuration object that was rebuilt many times and that have side effects by looking at proxymanager code
- if you dump the messaging configuration and warmup cache on your local computer, then move the cache to another folder, it will try to regenerate proxies in the the proxy path of your local computer (eg. /Users/jean/ecotone-dev/var/cache) instead of the folder where code is executed (eg. /app/var/cache). This is especially problematic in a lambda where you can write files only in the /tmp folder
With a bunch of lines, I think it is a lot easier to understand what is being done in a Gateway Proxy. I did not do this with performance in mind. I don't think this will have a noticeable impact on performance.
 
I will propose a PR with this change only but if we choose not to go this way, I will be happy to propose another PR the  autoloader registration/deregistration issue globally for Ecotone.